+++ src/Constraints/MultipleOf.php
@@ -2,28 +2,28 @@
 
 namespace League\JsonGuard\Constraints;
 
-use League\JsonGuard;
 use League\JsonGuard\Assert;
 use League\JsonGuard\ValidationError;
+use League\JsonGuard\Validator;
 
-class MultipleOf implements PropertyConstraint
+class MultipleOf implements Constraint
 {
     const KEYWORD = 'multipleOf';
 
     /**
      * {@inheritdoc}
      */
-    public static function validate($value, $multiple, $pointer = null)
+    public function validate($value, $parameter, Validator $validator)
     {
-        Assert::type($multiple, 'number', self::KEYWORD, $pointer);
-        Assert::nonNegative($multiple, self::KEYWORD, $pointer);
+        Assert::type($parameter, 'number', self::KEYWORD, $validator->getPointer());
+        Assert::nonNegative($parameter, self::KEYWORD, $validator->getPointer());
 
         if (!is_numeric($value)) {
             return null;
         }
 
         // for some reason fmod does not return 0 for cases like fmod(0.0075,0.0001) so I'm doing this manually.
-        $quotient = $value / $multiple;
+        $quotient = $value / $parameter;
         $mod      = $quotient - floor($quotient);
         if ($mod == 0) {
             return null;
@@ -33,8 +33,8 @@ class MultipleOf implements PropertyConstraint
             'Number {value} is not a multiple of {multiple_of}',
             self::KEYWORD,
             $value,
-            $pointer,
-            ['value' => $value, 'multiple_of' => $multiple]
+            $validator->getPointer(),
+            ['value' => $value, 'multiple_of' => $parameter]
         );
     }
 }