+++ src/Constraints/Pattern.php
@@ -5,23 +5,24 @@ namespace League\JsonGuard\Constraints;
 use League\JsonGuard;
 use League\JsonGuard\Assert;
 use League\JsonGuard\ValidationError;
+use League\JsonGuard\Validator;
 
-class Pattern implements PropertyConstraint
+class Pattern implements Constraint
 {
     const KEYWORD = 'pattern';
 
     /**
      * {@inheritdoc}
      */
-    public static function validate($value, $pattern, $pointer = null)
+    public function validate($value, $parameter, Validator $validator)
     {
-        Assert::type($pattern, 'string', self::KEYWORD, $pointer);
+        Assert::type($parameter, 'string', self::KEYWORD, $validator->getPointer());
 
         if (!is_string($value)) {
             return null;
         }
 
-        if (preg_match(JsonGuard\delimit_pattern($pattern), $value) === 1) {
+        if (preg_match(JsonGuard\delimit_pattern($parameter), $value) === 1) {
             return null;
         }
 
@@ -29,8 +30,8 @@ class Pattern implements PropertyConstraint
             'Value {value} does not match the pattern {pattern}.',
             self::KEYWORD,
             $value,
-            $pointer,
-            ['value' => $value, 'pattern' => $pattern]
+            $validator->getPointer(),
+            ['value' => $value, 'pattern' => $parameter]
         );
     }
 }