+++ tests/Constraints/TypeTest.php
@@ -4,6 +4,7 @@ namespace League\JsonGuard\Test\Constraints;
 
 use League\JsonGuard\Constraints\Type;
 use League\JsonGuard\ValidationError;
+use League\JsonGuard\Validator;
 
 class TypeTest extends \PHPUnit_Framework_TestCase
 {
@@ -11,7 +12,7 @@ class TypeTest extends \PHPUnit_Framework_TestCase
     {
         $type = new Type();
 
-        $error = $type->validate('1', 'number');
+        $error = $type->validate('1', 'number', new Validator([], new \stdClass()));
 
         $this->assertInstanceOf(ValidationError::class, $error);
     }